+++ evals/evals/marketplace/install-neon-postgres/EVAL.ts
@@ -1,33 +1,9 @@
 import { test, expect } from 'vitest';
-import { execSync, spawnSync } from 'child_process';
-import { readFileSync } from 'fs';
-
-function getToken(): string {
-  // Try reading from the Vercel CLI auth locations
-  const paths = [
-    '/home/vercel-sandbox/.local/share/com.vercel.cli/auth.json',
-    '/home/vercel-sandbox/.vercel/auth.json',
-  ];
-  for (const p of paths) {
-    try {
-      const auth = JSON.parse(readFileSync(p, 'utf-8'));
-      if (auth.token) return auth.token;
-    } catch {}
-  }
-
-  // Fall back to extracting from .bashrc
-  try {
-    const bashrc = readFileSync('/home/vercel-sandbox/.bashrc', 'utf-8');
-    const match = bashrc.match(/export VERCEL_TOKEN="([^"]+)"/);
-    if (match) return match[1];
-  } catch {}
-
-  throw new Error('No Vercel auth token found');
-}
+import { execSync } from 'child_process';
 
 function run(cmd: string): string {
-  const token = getToken();
-  return execSync(`${cmd} --token="${token}"`, {
+  // Source .bashrc to get VERCEL_TOKEN, then run the command
+  return execSync(`bash -c 'source ~/.bashrc 2>/dev/null; ${cmd}'`, {
     encoding: 'utf-8',
     stdio: 'pipe',
   }).trim();
