+++ packages/evals/README.md
@@ -7,7 +7,7 @@ Test AI coding agents to measure what actually works.
 1. **Install dependencies:**
 
    ```bash
-   npm install
+   pnpm install --ignore-workspace
    ```
 
 2. **Configure environment variables:**
@@ -21,6 +21,8 @@ Test AI coding agents to measure what actually works.
    - **Choose ONE agent key**: `AI_GATEWAY_API_KEY` (for Vercel agents), `ANTHROPIC_API_KEY`, or `OPENAI_API_KEY`
    - **Choose ONE sandbox option**: `VERCEL_TOKEN`, `VERCEL_OIDC_TOKEN`, or use Docker (set `sandbox: 'docker'` in config)
 
+   **Important:** Your `VERCEL_TOKEN` must be a Personal Access Token with access to a team/account where the evals can create and delete temporary test projects. Some evals (e.g. marketplace integrations) will provision resources against these projects during the run and clean them up afterwards.
+
 ## Running Evals
 
 ### Preview (no cost)
@@ -28,21 +30,15 @@ Test AI coding agents to measure what actually works.
 See what will run without making API calls:
 
 ```bash
-npx @vercel/agent-eval cc --dry
+npx @vercel/agent-eval vercel-cli-cc --dry
 ```
 
 ### Run Experiments
 
-Run the Claude Code experiment:
+Run the Vercel CLI marketplace eval:
 
 ```bash
-npx @vercel/agent-eval cc
-```
-
-Run the Codex experiment:
-
-```bash
-npx @vercel/agent-eval codex
+npx @vercel/agent-eval vercel-cli-cc
 ```
 
 ### View Results
